+++ app/Http/Controllers/UnitMeasureController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers;
 
 
 use App\UnitMeasure;
+use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Validation\ValidationException;
+use Symfony\Component\Translation\Exception\NotFoundResourceException;
 
 class UnitMeasureController extends Controller
 {
@@ -41,44 +43,55 @@ class UnitMeasureController extends Controller
                 ["errors" => $exception->validator->getMessageBag()],
                 JsonResponse::HTTP_UNPROCESSABLE_ENTITY);
         }
-        return response()->json(["UnitMeasure" => $unitMeasure]);
+        return response()->json(["UnitMeasure" => $unitMeasure],
+            JsonResponse::HTTP_CREATED);
 
 
     }
 
 
-    public function update(Request $request, UnitMeasure $unitMeasure)
+    public function update(Request $request, $id)
     {
         try {
-            $unitMeasure->validateAndFill($request->all());
-            $unitMeasure->save();
+            $unitMeasureSaved=UnitMeasure::find($id);
+            if(!$unitMeasureSaved)
+                throw new NotFoundResourceException("Unit Measure $id Not Found");
+            $unitMeasureSaved->validateAndFill($request->all());
+            $unitMeasureSaved->save();
         } catch (ValidationException $exception) {
 
             return response()->json(
                 ["errors" => $exception->validator->getMessageBag()],
                 JsonResponse::HTTP_UNPROCESSABLE_ENTITY);
         }
-        return response()->json(["UnitMeasure" => $unitMeasure]);
+        catch (NotFoundResourceException $notFoundE) {
+
+            return response()->json(
+                ["errors" => $notFoundE->getMessage()],
+                JsonResponse::HTTP_NOT_FOUND);
+        }
+        return response()->json(["UnitMeasure" => $unitMeasureSaved],
+            JsonResponse::HTTP_OK);
 
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param \App\Categories $categories
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(UnitMeasure $unitMeasure)
+
+
+    public function destroy($id)
     {
         //
         try {
+            $unitMeasure=UnitMeasure::find($id);
+            if(!$unitMeasure)
+                throw new NotFoundResourceException("Unit Measure $id Not Found");
             $unitMeasure->delete();
-        } catch (\Exception $exception) {
+        } catch (Exception $exception) {
 
             return response()->json(
                 ["errors" => $exception->getMessage()],
                 JsonResponse::HTTP_BAD_REQUEST);
         }
-        return response()->json(["UnitMeasure" => null]);
+        return response()->json(["UnitMeasure" => null],
+            JsonResponse::HTTP_NO_CONTENT);
     }
 }
